refact: aplicar responsabilidad unica a common
