+++ src/templates/flujos/common/datosinicialesComunes4.ts
@@ -1,0 +1,37 @@
+import { addKeyword, EVENTS } from '@builderbot/bot';
+import { consultarPacientePorDocumento } from '../../../services/apiService';
+import { consultarCitasPorDocumento } from '../../../utils/consultarCitasPorDocumento';
+import { obtenerCitasValidas } from '../../../utils/obtenerCitasValidas';
+import { datosinicialesComunes5 } from './datosinicialesComunes5';
+
+const datosinicialesComunes4 = addKeyword(EVENTS.ACTION)
+    .addAction(async (ctx, { state, flowDynamic, gotoFlow }) => {
+        const { tipoDoc, numeroDoc } = state.getMyState();
+        const paciente = await consultarPacientePorDocumento(numeroDoc);
+        if (!paciente || paciente.length === 0) {
+            await flowDynamic('No se encontró información del paciente con ese documento.');
+            return;
+        }
+        const primerNombre = paciente[0]?.PrimerNombre || '';
+        const segundoNombre = paciente[0]?.SegundoNombre || '';
+        const nombreCompleto = `${primerNombre} ${segundoNombre}`.trim();
+
+        const citas = await consultarCitasPorDocumento(tipoDoc, numeroDoc);
+        const ahora = new Date();
+        const citasValidas = await obtenerCitasValidas(citas, ahora);
+
+        await state.update({ citasProgramadas: citasValidas });
+        if (!citasValidas || citasValidas.length === 0) {
+            await flowDynamic('No se encontraron citas agendadas y vigentes con ese documento.');
+            return;
+        }
+        let mensaje = `Estimado/a *${nombreCompleto}* Tienes las siguientes citas agendadas y vigentes:\n`;
+        citasValidas.forEach((cita: any, idx: number) => {
+            mensaje += `*${idx + 1}*. *Fecha*: ${cita.FechaCita}, *Hora*: ${cita.HoraCita}, *Especialidad*: ${cita.Especialidad}\n`;
+        });
+        await flowDynamic(mensaje);
+        await state.update({ esperaSeleccionCita: true });
+        return gotoFlow(datosinicialesComunes5);
+    });
+
+export { datosinicialesComunes4 };
