+++ src/templates/flujos/common/index.ts
@@ -1,0 +1,6 @@
+export * from './datosInicialesComunes';
+export * from './datosinicialesComunes2';
+export * from './datosinicialesComunes3';
+export * from './datosinicialesComunes4';
+export * from './datosinicialesComunes5';
+export * from './volverMenuPrincipal';
