+++ src/utils/consultarCitasPorDocumento.ts
@@ -1,0 +1,18 @@
+import { consultarPacientePorDocumento, consultarCitasPorPacienteId } from '../services/apiService';
+
+export async function consultarCitasPorDocumento(tipoDoc: string, numeroDoc: string) {
+    const paciente = await consultarPacientePorDocumento(numeroDoc);
+    if (!paciente || paciente.length === 0) {
+        return [];
+    }
+    const pacienteId = paciente[0]?.PacientesID;
+    if (!pacienteId) {
+        return [];
+    }
+    const citas = await consultarCitasPorPacienteId(pacienteId);
+    if (!citas || citas.length === 0) {
+        return [];
+    }
+    const citasProgramadas = citas.filter((cita: any) => cita.EstadoAgenda === 'Programada');
+    return citasProgramadas;
+}
